# public/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Rest Api Slim PHP
  description: >
    [Example of REST API](https://github.com/maurobonfietti/rest-api-slim-php) with Slim PHP micro framework.

    This simple RESTful API made in Slim version 3, allows CRUD operations to manage entities like: Users, Tasks and Notes :-)
  contact: {}
  version: '1.0'
servers:
- url: https://rest-api-slim-php.herokuapp.com
  description: Production server using Heroku.

- url: http://localhost:8080
  description: My localhost server using PHP built-in web server.

- url: http://localhost:8081
  description: My localhost server using docker and docker compose.
paths:
...
  /api/v1/users:
    get:
      tags:
      - Users
      summary: Get All Users
      operationId: GetAllUsers
      parameters:
      - name: Authorization
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: '{{jwt}}'
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
      security: []
    post:
      tags:
      - Users
      summary: Create User
      description: Register a new user.
      operationId: CreateUser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            example:
              name: John User
              email: super.email@host.com
              password: OnePass1
        required: true
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
...
  /api/v1/tasks:
    get:
      tags:
      - Tasks
      summary: Get All Tasks
      description: Get all tasks of a user.
      operationId: GetAllTasks
      parameters:
      - name: Authorization
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: '{{jwt}}'
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
      security: []
    post:
      tags:
      - Tasks
      summary: Create Task
      description: Create a task.
      operationId: CreateTask
      parameters:
      - name: Authorization
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: '{{jwt}}'
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            example:
              name: Go To Sleep
              description: It's too late, go to sleep man ;-)
              status: 0
        required: true
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
      security: []
...
components:
  schemas:
...
    CreateUserRequest:
      title: CreateUserRequest
      required:
      - name
      - email
      - password
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      example:
        name: John User
        email: super.email@host.com
        password: OnePass1
...
    CreateTaskRequest:
      title: CreateTaskRequest
      required:
      - name
      - description
      - status
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: integer
          format: int32
      example:
        name: Go To Sleep
        description: It's too late, go to sleep man ;-)
        status: 0
...